fix exr memory allocation size overflow by turning the size into a uint64_t

+++ src/image_decode.cpp
@@ -547,7 +547,7 @@ namespace bimg
 						stepA  = 1;
 					}
 
-					data   = (uint8_t*)bx::alloc(_allocator, exrImage.width * exrImage.height * dstBpp/8);
+					data   = (uint8_t*)bx::alloc(_allocator, (size_t)exrImage.width * exrImage.height * dstBpp/8);
 					width  = exrImage.width;
 					height = exrImage.height;
